fixed `tensor` class
fixed broadcasting & operation functions
