+++ .gitignore
@@ -3,3 +3,6 @@
 __pycache__/
 *.py[cod]
 *.exe
+
+build
+.vscode
