--- axgrad/__init__.py
@@ -1,4 +1,0 @@
-from .tensor import *
-from .ops import *
-from .helpers.shape import flatten, broadcast_array
-from .helpers.utils import *
