+++ axon/__init__.py
@@ -1,4 +1,4 @@
-from .helpers.utils import ones, zeros, zeros_like, arange, randint, ns
+from .helpers.utils import ones, zeros, zeros_like, arange, randint
 from .tensor import tensor
 from .nn import *
 from .ops import matmul, stack, concat
