+++ axon/helpers/__init__.py
@@ -1,2 +1,2 @@
-from .utils import arange, ns, randint, zeros, zeros_like, randn
-from .shape import broadcast_array, check_arr, get_shape, view_arr, broadcast_shapes
+from .utils import *
+from .shape import broadcast_array, get_shape, broadcasted_shape
