+++ axon/helpers/utils.py
@@ -1,77 +1,14 @@
 import random
 
-def zeros_like(arr, dtype=int):
-  """
-    Create an array of zeros with the same shape and dtype as the input array.
-    Args:
-      arr (list or tuple): Input array.
-    Returns:
-      list: Array of zeros with the same shape and dtype as the input array.
-  """
-  if isinstance(arr, list):
-    return [zeros_like(elem) for elem in arr]
-  elif isinstance(arr, tuple):
-    return tuple(zeros_like(elem) for elem in arr)
-  else:
-    return dtype(0)
-
-def ones_like(arr, dtype=int):
-  """
-    Create an array of zeros with the same shape and dtype as the input array.
-    Args:
-      arr (list or tuple): Input array.
-    Returns:
-      list: Array of zeros with the same shape and dtype as the input array.
-  """
-  if isinstance(arr, list):
-    return [ones_like(elem) for elem in arr]
-  elif isinstance(arr, tuple):
-    return tuple(ones_like(elem) for elem in arr)
-  else:
-    return dtype(1)
-
-def zeros(shape, dtype=int):
-  """
-    creates an array of zeros according to the given input
-    Args:
-      shape (list or tuple): shape array.
-      dtype (int or float): determines the datatype
-    Returns:
-      list: array of zeros with the same shape as input_shape and in same dtype.
-  """
+def zeros(shape):
   if len(shape) == 1:
-    return [0] * shape[0] if dtype is None else [dtype(0)] * shape[0]
-  else:
-    return [zeros(shape[1:], dtype=dtype) for _ in range(shape[0])]
+    return [0] * shape[0]
+  return [zeros(shape[1:]) for _ in range(shape[0])]
 
-def ones(shape, dtype=int):
-  """
-    creates an array of ones according to the given input
-    Args:
-      shape (list or tuple): shape array.
-      dtype (int or float): determines the datatype
-    Returns:
-      list: array of ones with the same shape as input_shape and in same dtype.
-  """
+def ones(shape):
   if len(shape) == 1:
-    return [1] * shape[0] if dtype is None else [dtype(1)] * shape[0]
-  else:
-    return [ones(shape[1:], dtype=dtype) for _ in range(shape[0])]
-
-def ns(shape, n, dtype=int):
-  """
-    creates an array of 'n' according to the given input number
-    Args:
-      shape (list or tuple): shape array
-      n (int or float): sets the number for outputs
-      dtype (int or float): determines the datatype
-    Returns:
-      list: array of number equal to 'n' with the same shape as input_shape and in same dtype.
-  """
-  if len(shape) == 1:
-    return [n] * shape[0] if dtype is None else [dtype(n)] * shape[0]
-  else:
-    return [ns(shape[1:], dtype=dtype) for _ in range(shape[0])]
+    return [1] * shape[0]
+  return [ones(shape[1:]) for _ in range(shape[0])]
 
 def randint(low, high, size=None, dtype=int):
   if size is None:
@@ -86,13 +23,29 @@ def randint(low, high, size=None, dtype=int):
       return [dtype(random.randint(low, high)) for _ in range(size)]
 
 def arange(start, end, step):
-  return [start + i * step for i in range(int((end - start) / step))]
+  return [start + i * step for i in range(int((end-start)/step))]
 
 def randn(domain=(1, -1), shape=None):
   if len(shape) == 1:
     return [random.uniform(domain[0], domain[1]) for _ in range(shape[0])]
   else:
     return [randn(domain=domain, shape=shape[1:]) for _ in range(shape[0])]
+
+def zeros_like(arr, dtype=int):
+  if isinstance(arr, list):
+    return [zeros_like(elem) for elem in arr]
+  elif isinstance(arr, tuple):
+    return tuple(zeros_like(elem) for elem in arr)
+  else:
+    return dtype(0)
+
+def ones_like(arr, dtype=int):
+  if isinstance(arr, list):
+    return [ones_like(elem) for elem in arr]
+  elif isinstance(arr, tuple):
+    return tuple(ones_like(elem) for elem in arr)
+  else:
+    return dtype(1)
 
 def generate_random_list(shape):
   if len(shape) == 0:
